MAINT: Refactor energy decay runscript. Make model for it
